refuctoring code by controller
